+++ frontend/components/BookingCalendar.tsx
@@ -32,7 +32,7 @@ import {
 
 import { selectDatetime } from "../redux/datetimeSlice";
 import { useSelector } from "../redux/hooks";
-import RoomRatingList from "./Rooms/RoomRatingList";
+import RoomRatingList from "./Rating/RoomRatingList";
 
 const ToolBarButton = styled(Button)(({ theme }) => ({
   borderColor: theme.palette.secondary.main,
