+++ frontend/components/Rating/ReviewButton.tsx
@@ -1,0 +1,23 @@
+import { Button } from "@mui/material";
+import ReviewModal from "components/Rating/ReviewModal";
+import { useState } from "react";
+
+export default function ReviewButton() {
+  const [open, setOpen] = useState(false);
+
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
+
+  return (
+    <>
+      <Button
+        aria-label="leave-review-link"
+        onClick={handleOpen}
+        variant="outlined"
+      >
+        Leave a Rating
+      </Button>
+      <ReviewModal open={open} handleClose={handleClose} />
+    </>
+  );
+}
